% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordstatExtras-package.R
\docType{package}
\name{nordstatExtras-package}
\alias{nordstatExtras}
\alias{nordstatExtras-package}
\title{nordstatExtras: Shared SQLite Cache Backend for the nordstat Family}
\description{
Provides a SQLite-backed cell-level cache that can be used as a drop-in
backend by rKolada, rTrafa, and pixieweb. Designed for multi-user web
applications where minimal fetch latency and asynchronous writes are
required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LoveHansson/nordstatExtras}
  \item Report bugs at \url{https://github.com/LoveHansson/nordstatExtras/issues}
}

}
\author{
\strong{Maintainer}: Love Hansson \email{love.hansson@transportforetagen.se}

}
\keyword{internal}
